Run go work sync to sync the two go.mod files
Renovate update deps in the two go.mod files independently. So it can
miss the case when one go.mod update bumps an indirect dependency that
is also part of the other go.mod file for a different reason. This can
lead to divergent indirect dependencies in the two go.mod files.